requirements.txt
discord.py
vnstock3
pandas
google-generativeai
pytz
flask